add leader board slice - rejected csaes
