+++ packages/game-view/src/store/authSlice.test.ts
@@ -1,4 +1,4 @@
-import reducer, {
+import {
   loginWithEmailAndPassword,
   loginWithGitHubAuth,
   loginWithGoogleAuth,
@@ -9,99 +9,91 @@ import {
   signInWithGithub,
   signInWithGoogle,
 } from '../api/auth';
+import firebase from 'firebase';
 import { store } from './store';
 
 const sleep = (ms: number) => new Promise((res) => setTimeout(res, ms));
 
-describe('auth reducer tests', () => {
-  const initialState = { isAuth: true };
-
-  it('test if auth service is pending', () => {
-    const action = { type: loginWithEmailAndPassword.pending.type };
-    const state = reducer(initialState, action);
-    expect(state.isAuth).toBeFalsy();
-  });
-
-  it('test if auth service is resolved', () => {
-    const uid = nanoid(10);
-    const userName = nanoid(10);
-
-    const action = {
-      type: loginWithEmailAndPassword.fulfilled.type,
-      payload: { uid, userName },
-    };
-    const state = reducer(initialState, action);
-    expect(state).toEqual({ uid, isAuth: true, userName });
-  });
-
-  it('test if authService rejected', () => {
-    const action = {
-      type: loginWithEmailAndPassword.rejected.type,
-      error: { errorMessage: 'some error' },
-      payload: { error: 'some error' },
-    };
-    const state = reducer(initialState, action);
-    expect(state).toEqual({ isAuth: false });
-  });
-});
-
 jest.mock('../api/auth');
-describe('exampleSlice', () => {
-  it('reject thunk if auth service is error', async () => {
+describe('test login with email and pass - success', () => {
+  it('resolve thunk if auth service is success', async () => {
+    const rndUser = {
+      uid: nanoid(11),
+      displayName: 'aa', // nanoid(20),
+    };
     const doSignInWithEmailAndPasswordMock =
       doSignInWithEmailAndPassword as jest.MockedFunction<
         typeof doSignInWithEmailAndPassword
       >;
-    doSignInWithEmailAndPasswordMock.mockClear();
-    doSignInWithEmailAndPasswordMock.mockImplementation(() => {
-      throw new Error('some exception');
-    });
+    doSignInWithEmailAndPasswordMock.mockResolvedValue({
+      user: rndUser,
+    } as firebase.auth.UserCredential);
 
     store.dispatch(loginWithEmailAndPassword({ username: 'a', password: 'a' }));
     await sleep(0);
 
     expect(doSignInWithEmailAndPasswordMock).toBeCalledTimes(1);
+
     expect(store.getState().auth).toEqual({
-      errorMessage: 'some exception',
-      isAuth: false,
+      uid: rndUser.uid,
+      userName: rndUser.displayName,
+      isAuth: true,
     });
   });
 });
 
-describe('test login with googleAuth - Error', () => {
-  it('resolve thunk if auth service is Error', async () => {
+describe('test login with googleAuth - success', () => {
+  it('resolve thunk if auth service is success', async () => {
+    const rndUser = {
+      uid: nanoid(11),
+      photoUrl: nanoid(11),
+      displayName: nanoid(20),
+    };
     const signInWithGoogleMock = signInWithGoogle as jest.MockedFunction<
       typeof signInWithGoogle
     >;
-    signInWithGoogleMock.mockImplementation(() => {
-      throw new Error('some exception');
-    });
+    signInWithGoogleMock.mockResolvedValue(
+      Promise.resolve({
+        ...rndUser,
+      })
+    );
 
     store.dispatch(loginWithGoogleAuth());
     await sleep(0);
     expect(signInWithGoogle).toBeCalledTimes(1);
     expect(store.getState().auth).toEqual({
-      errorMessage: 'some exception',
-      isAuth: false,
+      uid: rndUser.uid,
+      userName: rndUser.displayName,
+      isAuth: true,
+      userPict: rndUser.photoUrl,
     });
   });
 });
 
-describe('test login with GitHubAuth - Error', () => {
-  it('reject GitHubAuth thunk if auth service is Error', async () => {
+describe('test login with GitHubAuth - success', () => {
+  it('resolve thunk if  GitHubAuth  service is success', async () => {
+    const rndUser = {
+      uid: nanoid(11),
+      photoUrl: nanoid(11),
+      displayName: nanoid(20),
+    };
     const signInWithGithubMock = signInWithGithub as jest.MockedFunction<
       typeof signInWithGithub
     >;
-    signInWithGithubMock.mockImplementation(() => {
-      throw new Error('some exception');
-    });
+    signInWithGithubMock.mockResolvedValue(
+      Promise.resolve({
+        ...rndUser,
+      })
+    );
 
     store.dispatch(loginWithGitHubAuth());
     await sleep(0);
-    expect(signInWithGithub).toBeCalledTimes(1);
+    expect(signInWithGithubMock).toBeCalledTimes(1);
     expect(store.getState().auth).toEqual({
-      errorMessage: 'some exception',
-      isAuth: false,
+      uid: rndUser.uid,
+      userName: rndUser.displayName,
+      isAuth: true,
+      userPict: rndUser.photoUrl,
     });
   });
 });
